+++ include/torch-mlir/Dialect/Torch/Utils/Utils.h
@@ -55,6 +55,12 @@ bool isViewLikeOp(Operation *op);
 Value getConstantWithGivenDtypeAndValue(PatternRewriter &rewriter, Location loc,
                                         float value, Type dtype);
 
+// Return the number of elements of a tensor if the shape is static; otherwise,
+// return -1.
+int64_t getNumberOfElements(RankedTensorType inputType);
+
+SmallVector<int64_t> makeShapeLLVMCompatible(ArrayRef<int64_t> shape);
+SmallVector<int64_t> makeShapeTorchCompatible(ArrayRef<int64_t> shape);
 } // namespace Torch
 } // namespace torch
 } // namespace mlir
